// src/shared/types/auth-status.ts
export enum AuthStatus {
    LoggedIn = 'LoggedIn',
    Auth = 'Auth',
    AuthError = 'AuthError',
    Register = 'Register',
    RegisterError = 'RegisterError',
    RegisterSuccess = 'RegisterSuccess',
    CheckEmailError = 'CheckEmailError',
    CheckEmailErrorNoExist = 'CheckEmailErrorNoExist',
    ConfirmEmail = 'ConfirmEmail',
    ChangePassword = 'ChangePassword',
    ChangePasswordError = 'ChangePasswordError',
    ChangePasswordSuccess = 'ChangePasswordSuccess'
}